Fix incorrect type conversion of token.

+++ rbxtestfiles/golden.go
@@ -435,7 +435,7 @@ func (g *Golden) Field(name string, v interface{}) *Golden {
 		g.popf(name)
 
 	case rbxfile.ValueToken:
-		g.uint(name, uint64(v))
+		g.uint(name, uint32(v))
 
 	case rbxfile.ValueReference:
 		g.ref("Value", v.Instance)
